+++ src/main/java/com/test/seminar/dao/impl/CourseDaoImpl.java
@@ -2,6 +2,8 @@ package com.test.seminar.dao.impl;
 
 import com.test.seminar.dao.CourseDao;
 import com.test.seminar.entity.Course;
+import com.test.seminar.exception.CourseNotFoundException;
+import com.test.seminar.exception.RepetitiveRecordException;
 import com.test.seminar.mapper.CourseMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -19,8 +21,12 @@ public class CourseDaoImpl implements CourseDao {
     CourseMapper courseMapper;
 
     @Override
-    public Course getCourseByCourseId(BigInteger courseId) {
-        return courseMapper.getCourseByCourseId(courseId);
+    public Course getCourseByCourseId(BigInteger courseId) throws CourseNotFoundException {
+        Course course = courseMapper.getCourseByCourseId(courseId);
+        if(course==null) {
+            throw new CourseNotFoundException();
+        }
+        return course;
     }
 
     @Override
@@ -34,17 +40,21 @@ public class CourseDaoImpl implements CourseDao {
     }
 
     @Override
-    public void insertCourse(Course course) {
+    public void insertCourse(Course course)throws RepetitiveRecordException {
         courseMapper.insertCourse(course);
     }
 
     @Override
-    public void updateCourseByCourseId(Course course) {
+    public void updateCourseByCourseId(Course course)throws CourseNotFoundException {
         courseMapper.updateCourseByCourseId(course);
     }
 
     @Override
-    public void deleteCourseByCourseId(BigInteger courseId) {
+    public void deleteCourseByCourseId(BigInteger courseId)throws CourseNotFoundException {
+        Course course = courseMapper.getCourseByCourseId(courseId);
+        if(course==null) {
+            throw new CourseNotFoundException();
+        }
         courseMapper.deleteCourseByCourseId(courseId);
     }
 }
