+++ src/main/java/com/test/seminar/dao/impl/RoundDaoImpl.java
@@ -2,6 +2,8 @@ package com.test.seminar.dao.impl;
 
 import com.test.seminar.dao.RoundDao;
 import com.test.seminar.entity.Round;
+import com.test.seminar.exception.RepetitiveRecordException;
+import com.test.seminar.exception.RoundNotFoundException;
 import com.test.seminar.mapper.RoundMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -19,23 +21,33 @@ public class RoundDaoImpl implements RoundDao {
     RoundMapper roundMapper;
 
     @Override
-    public void insertRound(Round round) {
+    public void insertRound(Round round)throws RepetitiveRecordException {
         roundMapper.insertRound(round);
     }
 
     @Override
-    public void deleteRoundByRoundId(BigInteger roundId) {
+    public void deleteRoundByRoundId(BigInteger roundId)throws RoundNotFoundException {
+        if(roundMapper.getRoundByRoundId(roundId)==null) {
+            throw new RoundNotFoundException();
+        }
         roundMapper.deleteRoundByRoundId(roundId);
     }
 
     @Override
-    public void updateRound(Round round) {
+    public void updateRound(Round round)throws RoundNotFoundException {
+        if(roundMapper.getRoundByRoundId(round.getId())==null) {
+            throw new RoundNotFoundException();
+        }
         roundMapper.updateRound(round);
     }
 
     @Override
-    public Round getRoundByRoundId(BigInteger roundId) {
-        return roundMapper.getRoundByRoundId(roundId);
+    public Round getRoundByRoundId(BigInteger roundId)throws RoundNotFoundException {
+        Round round=roundMapper.getRoundByRoundId(roundId);
+        if(round==null) {
+            throw new RoundNotFoundException();
+        }
+        return round;
     }
 
     @Override
