+++ src/main/java/com/test/seminar/dao/impl/TeamDaoImpl.java
@@ -2,6 +2,8 @@ package com.test.seminar.dao.impl;
 
 import com.test.seminar.dao.TeamDao;
 import com.test.seminar.entity.Team;
+import com.test.seminar.exception.RepetitiveRecordException;
+import com.test.seminar.exception.TeamNotFoundException;
 import com.test.seminar.mapper.TeamMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -19,28 +21,42 @@ public class TeamDaoImpl implements TeamDao {
     TeamMapper teamMapper;
 
     @Override
-    public Team getTeamByTeamId(BigInteger teamId) {
-        return teamMapper.getTeamByTeamId(teamId);
+    public Team getTeamByTeamId(BigInteger teamId)throws TeamNotFoundException {
+        Team team=teamMapper.getTeamByTeamId(teamId);
+        if(team==null) {
+            throw new TeamNotFoundException();
+        }
+        return team;
     }
 
     @Override
-    public void insertTeam(Team team) {
+    public void insertTeam(Team team)throws RepetitiveRecordException {
         teamMapper.insertTeam(team);
     }
 
     @Override
-    public void updateTeamByTeamId(Team team) {
+    public void updateTeamByTeamId(Team team)throws TeamNotFoundException {
+        if(teamMapper.getTeamByTeamId(team.getId())==null) {
+            throw new TeamNotFoundException();
+        }
         teamMapper.updateTeamByTeamId(team);
     }
 
     @Override
-    public void deleteTeamByTeamId(BigInteger teamId) {
+    public void deleteTeamByTeamId(BigInteger teamId)throws TeamNotFoundException {
+        if(teamMapper.getTeamByTeamId(teamId)==null) {
+            throw new TeamNotFoundException();
+        }
         teamMapper.deleteTeamByTeamId(teamId);
     }
 
     @Override
-    public Team getTeamByStudentIdAndCourseId(BigInteger studentId, BigInteger courseId) {
-        return teamMapper.getTeamByStudentIdAndCourseId(studentId, courseId);
+    public Team getTeamByStudentIdAndCourseId(BigInteger studentId, BigInteger courseId) throws TeamNotFoundException {
+        Team team=teamMapper.getTeamByStudentIdAndCourseId(studentId,courseId);
+        if(team==null) {
+            throw new TeamNotFoundException();
+        }
+        return team;
     }
 
     @Override
